# Conflicts:
#   backend/src/main/java/com/conkiri/domain/base/repository/ConcertRepository.java

+++ backend/src/main/java/com/conkiri/domain/base/repository/SeatRepository.java
@@ -10,5 +10,5 @@ import com.conkiri.domain.base.entity.Section;
 
 public interface SeatRepository extends JpaRepository<Seat, Long> {
 	List<Seat> findBySection(Section section);
-	Optional<Seat> findByRowLineAndColumnLineAndSection(Integer rowLine, Integer columnLine, Section section);
+	Optional<Seat> findByRowLineAndColumnLineAndSection(Long rowLine, Long columnLine, Section section);
 }
